+++ ui/main_window.py
@@ -106,6 +106,7 @@ class MainWindow(QMainWindow):
         self._tab_check.ti3_selected.connect(self._tab_profile.set_ti3_path)
         self._tab_check.ti2_found.connect(self._tab_measure.set_ti1_path)
         self._tab_print.ti2_loaded.connect(self._tab_measure.set_ti1_path)
+        self._tab_measure.ti2_loaded.connect(self._tab_print.apply_loaded_ti2)
         self._tab_print.ti2_replaced.connect(self._tab_profile.clear_files)
         self._tab_print.ti2_replaced.connect(self._tab_check.clear_files)
         self._tab_measure.ti2_replaced.connect(self._tab_profile.clear_files)